requirements.txt
llama-index
llama-index-embeddings-huggingface-api
llama-index-llms-groq
python-dotenv
fastapi[all]


llama-index-embeddings-huggingface-api